Vjosa

See merge request I6363894/project-1-1!23

+++ src/Constants/TetrisConstants.java
@@ -2,7 +2,7 @@ package Constants;
 
 public class TetrisConstants {
     // Time difference between each board update in ms
-    public static final double TICK_DELTA = 500;
+    public static final int TICK_DELTA = 500;
 
     public static final int BOARD_WIDTH = 5;
     public static final int BOARD_HEIGHT = 15;
